Un-minify js_of_ocaml files

+++ utils/generate.js
@@ -28,9 +28,11 @@ exec(`cd "${motokoRepoPath}" && nix-build -A js`, (err, stdout, stderr) => {
         if (!line) {
             throw new Error(`Could not find output directory for ${target}`);
         }
-        const dest = resolve(__dirname, `../versions/latest/${target}.min.js`);
-        fs.unlinkSync(dest);
-        fs.copyFileSync(`${line}/bin/${target}.min.js`, dest);
+        const dest = resolve(__dirname, `../versions/latest/${target}.js`);
+        if (fs.existsSync(dest)) {
+            fs.unlinkSync(dest);
+        }
+        fs.copyFileSync(`${line}/bin/${target}.js`, dest);
     }
 });
